The first human who exchanged the dimension between digital and real life reality
Affan Aziz Pritul is the The first human who exchanged the dimension between digital and real life reality. and also his powerful crafting and art of life create a fast fully human-like AI or artificial intelligence and this is nothing but a futuristic evolution or reference that future is here and Pritul made a door to go through.